I lied, heres the bluetooth object
